Optimise Route Planner and Fix up Dialogs
This optimises the route planner by only searching stars within a
bounding box - this speeds up finding paths thru the known star area
greatly. Also,
fixes so it remembers across invocations the last stars and range,
only allowing the find route button to be enabled when star data is
available and the systems entered in the dialog exist
Cleaned up output of dialog so easier to read.
Removed unnecessary duplication of data into other variables, the Graph
has all the data you need to added a few simple accessor functions to it
Graph needs computing each time as bounding box changes, overall
limiting stars is the most important to the speed as its limiting the
number of arcs is the important part to the overall speed.