Bugfix - fix duplicate entries in Trilateration View
We need to filter on system name as as comparing the cell.Tag reference no longer works now that we allow untrilaterated system.
This is because an untrilaterated system has a new SystemClass object created for it.

Alternatively duplicate the duplication check in getSystemForTrilateration() or similar earlier in the call chain.