Waypoint 6 - fix to visited system list.  Removed dead code from 3dmaps
Remove the animate dead code for now, and renamed _useTimer to
timerRunning to make it clear what it is doing..   Fix crash when double
clicking
Travelhistory ; split AddNewRow into two parts (same code).  One fixes
up the data, the second add new row to travelhistory
Always initialise all data in all entries in visitedsystemclass -
previously with a filter this was not happening.