Waypoint 5 - distance class and other fixes
1. Distance class - global removed, now uses functions in distanceclass
to find distance, using a SQL statement - combined both A&B and B&A in
one expression for speed.
2. Distance class used (thanks Anthor) id_edsm to keep in sync. New JSON
parser based on the one used for systems used to update DB.
3. Database has a new index for Distances:id_edsm and a new column for
it.  Clear down the EDSCLastDist value to force an update
4. SystemClass accepts a ISystem as its input to distance, and has a
distant function which asks the distance class for data - this
simplifies the other code.
5. Now commit 100000 items at a time
6. VisitedSystemsClass is now told if DB distances should be used during
update.
7. EDDBClass Distances are now downloaded en mass from Anthors server,
not Finwen old copy.
8. Discovery form: Removed redundant code,
galactic mapping downloaded once per 7 days, not held in db, held in
memory.
EDSM distance code changed to use new distance functions.  Only calls
OnDistancesLoaded if a change occurs to the distance database - this
prevents the refresh when not using distances or nothing changed
Move verbose about what is happening
9. EDSMClass request distances now asks for the showid.  File is
downloaded from edsm.
10. 3dmaptext updated to tell about shift
11. Saverouteexpedition uses systemclass DistanceIncludeDB
12. Setting text on distants tells user its slower and not needed
13. Travel history control more verbose, selects ALL as its default
travel filter, passes if DB is to be used for distances, and prevents
the refresh when the commander box is filled in (this would cause the
screen to not appear for a long time).
14. A lot.