Waypoint 10 - Restructure of the sync process with EDSM/EDDB
One sync thread now for both, because we need to sync EDSM first then
EDDB, then distances, and then refresh the history.   Use bool flags to
tell this function what to do and allow it to be kicked in by the user.
This means 1 EDSM and EDDB can't clash with db access, 2 EDDB is done
after EDSM is happy with the data