Waypoint 15 - SQLite data base cleaned up
1. Now has a CreateConnection which everyone uses - returns a
SQLiteConnection.  This is where the dB check is done.
2. All use CreateCommand - this cleans up the code considerably, and
ensure we have a consistent pattern, sets timeout.
2. Cleaned up db create logic - now in one function, open/close inline,
much clearer to read, not performing multiple upgrade calls etc.
3. All SQLiteDBClass gets/puts support either call with no connection or
with connection
4. All SQLiteSQLQueries support call with open or closed connection, and
leaves it that way
5. All SQLiteDBClass puts are optimised to open only the db once.
6. All other DB classes now use this harness, considerably cleaning up
the code, and also optimising the number of opens.  The Adds save 1
open.  There are also other examples where the number of opens have been
cut.