Waypoint 17 - minor improvements
1. Region bookmarks are in yellow, which was missed
2. Optimised start up EDSM sync - if no systems are in the table, don't
do the sync lookups. Speeds up this part of picking up when newly
installed.
3. Use LogLine throughout EDDiscoveryforms
4. On first run, present warning dialogs to user that it will take time
to sync, and a prompt to reboot after the first sync.
I tried to update SystemNames after load but as expected it causes a
hang up.. It hates doing this after its been assigned to a control.
5. AboutBox - stop repeating the version number..