Netlog class thread safe
Travelhistorycontrol newposition now called in foreground.  ParseFiles
now stops monitor for you - you use need to restart it.  Passes back
error string to print

Netlogclass:
1) does not need any lock statements, making sure that the monitor is
off during the foreground refresh parsefiles
2) does the visiting systems update in the foreground thread so we don't
update visited systems behind the back of the UI thread
3) Allow multiple system updates per parse, in case it misses something
4) simplifies onnewposition, since its always called in the foreground
5) ensure that visiting system is fully filled, before the UI thread
gets to it and starts using its data (that was a bug)
6) Checked against race conditions where a new file is made, but is
empty, but the thread gets to it before a write occurs
7) Does not run two monitor threads on top of each other
8) ensures that the last file written is the one monitored at the start.