Improve naming performance
If name is there, just schedule a newvertices update.  This means
texture buffer is not loaded, cutting down time to update.  Also limit
the updates to  200 per go, and ask for a repaint if we need more.
Performance on my machine is smooth.  Only the clean up makes it jump,
which happens infrequently.