Waypoint 5 Working with new resmat, investigating performance
Now uses a computed resmat by us, instead of the open GL.  This means
the code can have the resmat as soon as the position is updated, and
then this feeds into the star naming system immediately. Solves the
problem of Sol not popping up

Added debug features to try and work out optimal timing of tick/paint -
added a simple debug output in tools to help.

The resmat for non prespective mode may not be perfect.

Fix problem with systemclass crashing if no stars are found but its Sol
grid