Use a queue between the star name estimator and the draw to prevent lock
Estimator now not blocked, instead queues up the starts and the draw
removes them into the dictionary.

The requires sequencing of estimate-est complete-draw using some bools
in the form

Would be nice to remove the dictionary in the estimator.. then it would
never need to block, just feed new stars to the foreground.  but that
would mean it would probably over compute.. need to think