Waypoint 7 Fix to starnaming and perf optimisations, FPS counter
CDM fixed names and added Any func. Now forces thru a move
FPS counter now turned on/off by ctrl+F.  Fixed keyhander to cope.
F1 is help
F3/F4 is star name limit which is working again.
Remove dependency in keyhandler on perspective mode - now this is dealt
with in the correct place in positiondirection
To stop gitter, only perform a star rename if on a tick the grids did
not update
Status bar only updated on movement, not all the time, unless FPS
counter is on
Star name list : Remember zoom/rotation of painted image and use that to
determine if a redraw is needed.  This means turning around quickly on
an already drawn image works
Star name list: make sure we store the location in vector3 in same
format as inview list so it can be matched in the dictionary properly
Reduce number of stars per draw.