Added code to set EDSMId and add JSON for position when EDSM ID looked up
Logic lazy loads and matches edsm id.. if a star has a co-ord in json,
its happy.  then the front end will do the EDSM match only when it needs
to.  It will do this on assigned edsm_id, then co-ord or name.
If a star does not have co-ord, its is found by name, and the co-ord is
added back to the json so it won't penalise the start up next time