Switch the DB connection locks to reader-writer locks
Use a ReaderWriterLock instead of a Monitor when locking
the database, in order to allow multiple threads to read.
Also reduce the lifetime of some connections to that
of the inner transaction.
This should clear up some of the performance issues with
DB access, and further prevent any "database is locked"
errors.