Routing information should be read-only
If only I could learn to stop trying to pilot my ship while I've got my
next destination selected for copying in EDD. At least with this patch,
the destination won't be aaadadddadaddadadwj.

RichTextBoxScroll has the following line in reference to this:
	public RichTextBoxBack TextBox;  // Use these with caution
Not sure where a problem could arise from using this, but I wanted to note
it here. I'm also not very familiar with the EDD code base, so YMMV.