Add a checkbox to enable the new system tray icon
Not that anyone will want to disable it, but I guess that some people are
purists. Defaults to unchecked, in compliance with usability
recommendations from freedesktop. Of course, very few people will even
be running this on a Linux system, so freedesktop probably isn't the right
bar to be aiming for.

Also, fixup this groupbox's tab order.