Using CellEnter would cause a UI crash if cursor keys held down
Due to cell enter being called by the underlying system when the grid
row was filled, and the tag not being set up yet.

So, don't use cell enter (it was a clodge anyway to get the cursor keys
working).  Instead use key down/up and use that to decide when to
update.