AppContext: let base Dispose MainForm naturally
MainForm is technically ApplicationContext's problem, and it'll dispose
  of it appropriately without our help.
Fix an NRE if EDDiscoveryForm.ctor is having a bad day (~missing dlls)

TODO: We need an exception dialog. Something that'll handle whatever is
thrown at it, be it an Exception, an Exception.InnerException, even an
ExceptionCollection or AggregateException. This dialog should behave
similar to the default debug AppDomain.UnhandledException handler.