AutoCompleteDGVColumn Solidification
An abstract AutoCompleteDGVColumn isn't effective if the inheriting
classes are identical to base in all but type name. Remove the abstract
qualifier from it, and remove the essentially identical child types.

Cleanup some IDE0019 (Use pattern matching) messages.

Remove a couple of `using EDDiscovery;` lines from types that can't be
linked to EDDiscovery.