Mainform dblclick in non-client toggles maximize
Some clever Timer usage, and a cached MouseEventArgs, allows us to
toggle maximized state when double-clicking on a caption-esque control.

NOTE: this changes the behavior of of the caption controls to only
handle mousedown when it's the left mouse button and the form is drawn
without the window frame. This may not neccessarily be desired, but is,
IMHO, the correct thing to do.