Fixed back up pop out control so disassociates display order from save order, made tab strip nicer
We still save the enum in the registry for which tab is selected
(backwards compatible).  Fixed incorrect insertion of enum for
EstimatedValues and made notes NOT TO REORDER THEM!

To allow presentation order to be changed, rewrote how PopOutList works
in pop out control - its a list, with the presentation order, and can be
different from the enum.  Worked thru the functions to take either an
open index or enum to open a panel.
Tabstrip has autorepeat on its left/right arrows.