Fixed up EDSM lookup of system more
SystemCache returns null if it can't find anything, as the previous
function before system cache was introduced did.  This is a signal not
to keep on trying.  Removed unused func
HistoryEntry, if we tried to find a system without co-ords, but failed,
mark it as a failure in newsys.id_edsm so the front end system does not
try again
HistoryList:
Fixed up FillinPositionsFSDJumps more to check correctly before
attempting a system lookup.. mostly it should always do nothing as the
historylist accumulate should have dealt with it
FillEDSM - no reload function anymore (pointless, unlikely a new system
will have appeared and it just wastes lookup time)
Now correctly marks the entries as -1 if it can't find the system from
the cache.