Removed IDiscoveryController and simply RequestHistoryAsync from point of EDD form
EDD form has simple RefreshAsync, to simplfy the use case thru it.  If
it wants to change commander, it uses the controller directly.
Discoveryform does not need to be a controller.

Controller - more comments on how this whole system works around
refresh/edsm.

Added a debug to systemclassdb for assisting finding dups