Sort out Sort across all user controls PR1512
PR1512 fixed the sorting and memory of sorting, and cleaned up the
implementation of Stats. Cleaned up the visuals of the user control in
stats and make it clean.
And then cleaned up all sorting, inclduigng removing the old
DataGridViewSorter and converting all UCs to use SortCompare, and show
visual tick to indicate sort.
Added some baseutils to assist in sorting.