ShipYard and Remove multiple entries when reading journal
Shipyard introduced to find ships - new UC.  Icon not ready yet but will
be soon
Classes moved to first class which are used in more than just the
journal entry
Reader now screens out multiple repeats of shipyard, storedships,stored
modules, outfitting, market so that we don't blam the journal with
repeated entries.  On undock, all repeats are cleared.
Change to form to make sure the TLU gets updated even if we don't create
a JR.. this was a fault, and caused re-reading of old entries in the
journal if you restarted.