Fix Reported problems with Sold Ships
SellOldShip not passing correct FD ID to ship info list so was not being deleted for some types.
Reworked display of Ships in loadouts/ships to put SRVs at the bottom
Changed Sold in Shipinfo to State - owned, destroyed, sold.. this is for the future..  at the moment only owned/sold is used.
Hooked up DestroyedSRV and FighterDestroyed for the future into SIList.