Fix issue with ModuleBuy and SRVs/Fighters
Module buy was being called on a taipan, and creating a null entry in the list without a fdname.  This meant other bits would fail much later.

Loadgame does not call shipinfo if it does not have a valid ship name

Modules buy etc do not assume you are in the ship - you may be buying for a SRV/Fighter.