Fix compass and bookmark #2045
Surface bookmarks - if a planet is not in list, add it in, so we don't get cell validation errors.  This allows for a planet mark to show even if the planet list does not currently have that planet.
Allow editing of location name
Compass - current bookmark was not being set on display - so clicking on bookmark button make it make a new one not an updated one.  Fixed.  Disable bookmark button until last_he is set so it does not crash