Move history entry status to separate class
This should reduce the amount of memory each history entry consumes -
one object reference per history entry, plus the history status entry
for each status change (location / dock / undock / crew / etc).  In
practice, with a history of 41k entries with about 13k jumps, it reduces
the memory consumption by about 1.1MB.

It also separates Body from Station, so `WhereAmI` should be the station
name when docked and the body name (which can still be the station if
it's an orbital station) when undocked.  It should also reset to the
system name when supercruise is engaged away from an orbital station.