Pushed Journal 3.3 for ActiveStates FSD Location, Happiness, and fixed journal reading delay
During reading journals with no additional files, the wait period for it to appear was being triggered, meaning it could not read them.

Push the fact that we are reading historical records to the read additional files and only pend if we are reading them real time.

Added happiness

Fixed bodytype - normalised null to barycentre as per journal.