Rework FillSummaryName
he.EventSummary stays the same, but the Journal FillSummary changes to take a system parameter.

No caching of eventsummaryname in class - always use the dictionary to look it up. The caching was before the dictionary.
Passing system thru SummaryName() now allows the overrides to customise on system,

Scan does this to shorten the name of the planet in the summary so you can see it

Also added some scan types descriptions to the summary name so you know what kind of scan it is.

Made the full name appear all times so you can still search for all scans on it in travel grid

distance ls given in normal view, since its important.