Simplified the calls for EstimatedValue.
Pushed value of mapped/efficiency to the scan journal - this makes it easy to now compute all three estimated values - normal+mapped, first discovered, first mapped..

Removed lots of code in the users of estimated values

This is present on the text list for now.

Scan node re-orgnanised, its a bit of a long class. Kept the remembering of mapped/efficiency in there since scans can be replaced.  Pushed down the value of that to the scan node when its set, or SAA Scan is done.