Allow Journal to create UIEvents
This is to allow the journal to create sendtext (as command)/music events which do not go into the journal or the db.  the mechanism is now in there to do this.

Rework the music event so its always a UI event - now removed from db
Add the sendtext /r as a command and rely as a UI event - thats removed from db. normal sendtext is not

Config and settings lose the UI show music bit from the journals
UIEvent interface now removed.
Historylist always removes historic music events from the HL - even though they will be in the db they are not shown.
Journal scan code now can produce UI and journal events

SPanel also now uses GUIFocus to determine show and it works