Rework the EDSM/EDDB download to use new downloader and delete after them
Reducing footprint
EDDB downloads to root folder instead of a special one and cleans up afterwards
EDSM downloads and cleans up.  It downloads the whole file first on purpose - this is because if the user cancels it won't need to do the whole download again.
All use the new downloader
If initial EDSM download fails, it does not try the slow update