#2619 Fix galaxy selector - broken and slow down deletes with gaps
During DB change, .ID() in galaxy selector was calling the wrong function (the int/128)

To prevent this, change the name of the function to  ID128 and make sure the galaxy selector is using the double one.

This was preventing galaxy selection.

Also slow it down with gaps so it does not hog the db