#2858 Grid Alt Colour, and #2958 backup option
Backup option on safe mode form
Grid alt colour implemented. Note this means travel grid is now in standard colours, not visited colour. If a system without co-ords is present the short form is in red.
TG has a right click to restore the visited colour if you so want it.