Fix Route, Fix Surveyor
GMO no longer holds the enable, the Form3D does.  Preparation for the future
Surveyor now works again - star scan system can accept a Isystem with just a name, but surveyor during play now synthesises a ISystem with sys addr and name
Route rewritten how it operates to remove the requirement to do a FindSystem on power on.  Simpler logic.