Use new Registry, normalise Get/Put access to DB in all UCs
Use the new UCCB Get/Put functions instead of the direct DB access
Normalise the naming- 98% same names, a few are changed due to the slightly random names used before.
Most of this is grep/replace, i've checked the panels for naming as I converted them.

I suspect a few panels were not working with profiling properly so this should fix it.

This is in preparation maybe for a grand rename.