Initial CAPI, fixes to history panels for updates
CAPI login done, can now login using settings | commander | login.  CAPI buttons react to login

Experimental recode of journal, travel, star list to remove loadcomplete (race condition on timer), and recode the add so it does not rely on displayedrowcount but selected the row 0 on cursor without this.  This may solve the issue with the travelgrid not updating if not selected