Fix issue with transparency, and with & in window titles
Form was not working nicely with windows borders, and not displaying & due to UseMnemonic in labels.

Also made it play nicer with border on showing the top control line when hovered over.

Fixed issue with positioning of UCCB in panel if the top line is on and window border is on

Renamed some interfaces for transparency because they are confusing what they mean - IsTransparentOn -> IsTransparentModeOn