#3478 fix and enhance so more tabs can go to the JID on the travel grid, and allow it to be made visible
Remove Requestpanelaction long (JID) and introduce RequestTravelToJID class, with a MakeVisible member to allow the panel to show after selection. Return if it succeeds

UCCB has a ParentUCCB member, used in search to point to its parent uccb and the sub panels can use these.  This fixes 3478 assert

Datagridstarresults only shows goto if the handler is installed

Error message if grid does not have entry due to filtering