Restore nuget packages for submodule solutions
When a non-SDK project references a package, it will reference it in the
packages directory in the solution root. As a result, if a package is
updated within the submodule solution, it will reference the packages in
the root of that solution rather than in the root of the including
solution.